+++ src/finlang.test/TranspilerTest/C99TranspilerTest.cs
@@ -1,4 +1,4 @@
-using finlang.test.TranspilerTest;
+﻿using finlang.test.TranspilerTest;
 
 namespace finlang.TranspilerTest;
 
@@ -32,6 +32,60 @@ public class C99TranspilerTest
     //    files.HasFileName("MyClass.h").Should().BeTrue();
     //    files.HasFileName("MyClass.c").Should().BeFalse();
     //}
+
+    /// <summary>
+    /// https://github.com/fin-language/fin/issues/80
+    /// </summary>
+    [Fact]
+    public void MinMaxConstants()
+    {
+        const string code = """
+            using finlang;
+
+            public class MyClass: FinObj
+            {
+                static u8 get_u8_max() {  return u8.MAX /*u8.MAX*/;  }
+                static u16 get_u16_max() {  return u16.MAX /*u16.MAX*/;  }
+                static u32 get_u32_max() {  return u32.MAX /*u32.MAX*/;  }
+                static u64 get_u64_max() {  return u64.MAX /*u64.MAX*/;  }
+            
+                static i8 get_i8_max() {  return i8.MAX /*i8.MAX*/;  }
+                static i16 get_i16_max() {  return i16.MAX /*i16.MAX*/;  }
+                static i32 get_i32_max() {  return i32.MAX /*i32.MAX*/;  }
+                static i64 get_i64_max() {  return i64.MAX /*i64.MAX*/;  }
+
+                static u8 get_u8_min() {  return u8.MIN /*u8.MIN*/;  }
+                static u16 get_u16_min() {  return u16.MIN /*u16.MIN*/;  }
+                static u32 get_u32_min() {  return u32.MIN /*u32.MIN*/;  }
+                static u64 get_u64_min() {  return u64.MIN /*u64.MIN*/;  }
+
+                static i8 get_i8_min() {  return i8.MIN /*i8.MIN*/;  }
+                static i16 get_i16_min() {  return i16.MIN /*i16.MIN*/;  }
+                static i32 get_i32_min() {  return i32.MIN /*i32.MIN*/;  }
+                static i64 get_i64_min() {  return i64.MIN /*i64.MIN*/;  }
+            }
+            """;
+
+        var files = TranspilerTestHelper.TranspileFinToCFilesWithDummyMain(code);
+        string generatedCode = files.GetSingleWriterTextByFileName("MyClass.c");
+        generatedCode.Should().Contain("return UINT8_MAX /*u8.MAX*/;");
+        generatedCode.Should().Contain("return UINT16_MAX /*u16.MAX*/;");
+        generatedCode.Should().Contain("return UINT32_MAX /*u32.MAX*/;");
+        generatedCode.Should().Contain("return UINT64_MAX /*u64.MAX*/;");
+        generatedCode.Should().Contain("return INT8_MAX /*i8.MAX*/;");
+        generatedCode.Should().Contain("return INT16_MAX /*i16.MAX*/;");
+        generatedCode.Should().Contain("return INT32_MAX /*i32.MAX*/;");
+        generatedCode.Should().Contain("return INT64_MAX /*i64.MAX*/;");
+
+        generatedCode.Should().Contain("return 0 /*u8.MIN*/;");
+        generatedCode.Should().Contain("return 0 /*u16.MIN*/;");
+        generatedCode.Should().Contain("return 0 /*u32.MIN*/;");
+        generatedCode.Should().Contain("return 0 /*u64.MIN*/;");
+        generatedCode.Should().Contain("return INT8_MIN /*i8.MIN*/;");
+        generatedCode.Should().Contain("return INT16_MIN /*i16.MIN*/;");
+        generatedCode.Should().Contain("return INT32_MIN /*i32.MIN*/;");
+        generatedCode.Should().Contain("return INT64_MIN /*i64.MIN*/;");
+    }
 
     /// <summary>
     /// https://github.com/fin-language/fin/issues/30