+++ src/test/ExSln2/LedBlinker/test_stuff/NumericsMinMaxConstants.cs
@@ -1,0 +1,14 @@
+﻿using finlang;
+
+namespace hal;
+
+/// <summary>
+/// https://github.com/fin-language/fin/issues/80
+/// </summary>
+public class NumericsMinMaxConstants : FinObj
+{
+    public static u8 get_max()
+    {
+        return u8.MAX;
+    }
+}