+++ src/test/ExSln2/c99/gen/hal_CArraySizedEx.h
@@ -9,6 +9,7 @@
 
 
 
+
 typedef struct hal_CArraySizedEx hal_CArraySizedEx;
 struct hal_CArraySizedEx
 {